fix(packaging): moving root test classes to sampleparserjavaproject package
for consistency
