+++ src/test/kotlin/com/appdetex/sampleparserjavaproject/GooglePlayAppTest.kt
@@ -1,9 +1,7 @@
-package com.appdetex
+package com.appdetex.sampleparserjavaproject
 
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
-
-import com.appdetex.sampleparserjavaproject.GooglePlayApp
 
 class GooglePlayAppTest : StringSpec({
     val title = "Minecraft"
