+++ src/test/kotlin/com/appdetex/sampleparserjavaproject/MainTest.kt
@@ -1,4 +1,4 @@
-package com.appdetex
+package com.appdetex.sampleparserjavaproject
 
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
